add tool.py
